Add files via upload
Built pitch type prediction model. Basic idea is logistic model with mean for player, league wide adjustment for specific handedness/counts, and then player specific linear betas to strikes/balls/handedness. Prior for the means was a mixture of two normals with the idea that pitchers either have the pitch in their arsenal or not. I did leave it so that the posteriors weren't exactly 0 but generally under 1/500 or so if they didn't pitch it in the dataset.